automarker/SQAM/requirements.txt
python-Levenshtein
fuzzywuzzy
textdistance
statistics
mysql-connector-python
py-postgresql
#psycopg2
psycopg2-binary
jinja2
aio_pika
sqlparse